+++ rules/index.ts
@@ -1,6 +1,6 @@
-export { default as noBooleanParams } from './no-boolean-params.js';
-export { default as noMagicNumbersExceptZeroOne } from './no-magic-numbers-except-zero-one.js';
-export { default as enforceMeaningfulNames } from './enforce-meaningful-names.js';
-export { default as noLateArgumentUsage } from './no-late-argument-usage.js';
-export { default as noLateVariableUsage } from './no-late-variable-usage.js';
-export { default as lowFunctionCohesion } from './low-function-cohesion.js'; 
+export { default as noBooleanParams } from './no-boolean-params';
+export { default as noMagicNumbersExceptZeroOne } from './no-magic-numbers-except-zero-one';
+export { default as enforceMeaningfulNames } from './enforce-meaningful-names';
+export { default as noLateArgumentUsage } from './no-late-argument-usage';
+export { default as noLateVariableUsage } from './no-late-variable-usage';
+export { default as lowFunctionCohesion } from './low-function-cohesion';