+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2024-09-15
 
 ### Added
-- Initial release of eslint-plugin-my-rules
+- Initial release of eslint-plugin-code-complete
 - Rule: no-boolean-params
 - Rule: no-magic-numbers-except-zero-one
 - Rule: enforce-meaningful-names